functions/requirements.txt
functions-framework==3.*
google-api-python-client
google-auth
pandas
